+++ lib/forked/parse/parse_generic.rb
@@ -1,0 +1,10 @@
+module Forked
+  # Forked story file parser
+  class Parser
+    class << self
+      def parse_generic(*rgs)
+        return nil
+      end
+    end
+  end
+end
