+++ lib/forked/parse/parse_paragraph.rb
